Find Shattered Islands operator, Updated PR
Since last PR:
I renamed the operator to Find Shattered Islands (previously -UVs)
I updated the icon to be a hybrid of my designs (B and C), I simply call it the new B. The SVG is included in the files.
The operator's Minimum Acceptance property was changed into a logically inclusive value simply called Threshold for improved clarity. The property was also given a min value of 1.